feat: display feed certification in discover result
